web0/requirements.txt
flask-restful
redis
requests
requests_oauthlib
simplejson
pybloom_live
schedule
pyopenssl
